pyproject.toml
[project]
name = "master-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.1.3",
    "langchain-deepseek>=1.0.1",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
]
